Add Authentication Credentials (#3)
This PR adds a `Credentials` struct which implements the
`credentials.PerRPCCredentials` interface in order to authenticate
requests to Ensign via Quarterdeck. It also adds a Quarterdeck client
that can `Authenticate` and `Refresh` with a client ID and secret. This
is a partial implementation of the Quarterdeck API client in the
`ensign` repo that only needs to do machine authentication in the SDK
and reduces circular dependencies.